requirements.txt
fastapi
uvicorn
python-multipart
openai
pytesseract
Pillow
PyMuPDF
